# backend/docs/swagger.yaml
basePath: /
definitions:
  models.Admin:
    properties:
      email:
        type: string
      employees:
        items:
          $ref: '#/definitions/models.User'
        type: array
      id:
        type: string
      name:
        type: string
    type: object
  models.User:
    properties:
      admin_id:
        type: string
      code:
        type: integer
      id:
        type: string
      name:
        type: string
    type: object
  schemas.CreateAdminInput:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    required:
    - email
    - name
    - password
    type: object
  schemas.CreateUser:
    properties:
      admin_id:
        type: string
      name:
        type: string
    required:
    - admin_id
    - name
    type: object
  schemas.LoginByCode:
    properties:
      code:
        type: integer
    required:
    - code
    type: object
host: localhost:8080
info:
  contact: {}
  description: Сервис трекинга на Golang.
  termsOfService: http://swagger.io/terms/
  title: GoTracker API
  version: "1.0"
paths:
  /admin:
    post:
      consumes:
      - application/json
      description: Регистрирует администратора, хеширует его пароль и сохраняет в
        БД
      parameters:
      - description: Данные для создания админа
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/schemas.CreateAdminInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Admin'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Создать нового администратора
      tags:
      - admin
  /admin/message:
    post:
      consumes:
      - application/json
      description: Проверяет, является ли пользователь подчиненным (коллаборатором)
        данного администратора, и отправляет ему WebSocket-уведомление SHOW_DUCK с
        сообщением.
      parameters:
      - description: Данные для триггера утки (admin_id, user_id, message)
        in: body
        name: input
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Уведомление успешно отправлено
          schema:
            type: string
        "400":
          description: Неверный формат JSON или отсутствуют обязательные поля
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Доступ запрещен (пользователь не является коллаборатором этого
            админа)
          schema:
            additionalProperties:
              type: string
            type: object
        "405":
          description: Метод не поддерживается (разрешен только POST)
          schema:
            type: string
      summary: Отправка уведомления пользователю от администратора
      tags:
      - admin
  /ping:
    get:
      description: Возвращает текстовый ответ "pong", если сервис успешно запущен
        и принимает запросы
      produces:
      - text/plain
      responses:
        "200":
          description: pong
          schema:
            type: string
        "405":
          description: Method Not Allowed
          schema:
            type: string
      summary: Проверка работоспособности (Healthcheck)
      tags:
      - system
  /users:
    get:
      description: Возвращает список всех пользователей из базы данных
      produces:
      - application/json
      responses:
        "200":
          description: Успешный ответ со списком пользователей
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить всех пользователей
      tags:
      - users
  /users/:
    post:
      consumes:
      - application/json
      description: Принимает данные нового пользователя, привязывает к существующему
        админу и возвращает созданный объект
      parameters:
      - description: Данные пользователя (передай name и admin_id)
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/schemas.CreateUser'
      produces:
      - application/json
      responses:
        "201":
          description: Пользователь успешно создан
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Неверный формат JSON или админ не найден
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Создание пользователя админом
      tags:
      - users
  /users/login:
    post:
      consumes:
      - application/json
      description: Принимает код входа, проверяет его и возвращает данные пользователя
        для последующего WS-подключения
      parameters:
      - description: Код входа пользователя
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/schemas.LoginByCode'
      produces:
      - application/json
      responses:
        "200":
          description: Пользователь успешно вошёл
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Неверный формат JSON или пользователь не найден
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Вход пользователя по коду от администратора
      tags:
      - users
swagger: "2.0"